🎨 restructure providers
